policies/mcp_bot/requirements.txt
pyyaml>=6.0
requests>=2.31.0
certifi>=2023.0.0
google-genai>=0.2.0
PyGithub>=2.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

